Move Implicit-specific types to Features/ImplicitFlow (#90)
The Implicit-specific response processors (TokenProcessor,
IdTokenProcessor) and the EnableImplicitFlow() opt-in extension
move under Abblix.Oidc.Server/Features/ImplicitFlow, alongside
the other feature folders (PushedAuthorization, BackChannel-
Authentication, SessionManagement, etc.). The shared
IAuthorizationResponseProcessor abstraction and the default
AuthorizationCodeProcessor stay in Endpoints/Authorization
because they are not Implicit-specific — they are the protocol
the orchestrator dispatches over for every flow.

Hosts that opt into Implicit Flow now write
`using Abblix.Oidc.Server.Features.ImplicitFlow;` and
`services.EnableImplicitFlow()`, matching the convention used
for other opt-in features in this repo.

No behavior change.
