+++ Abblix.Oidc.Server/Endpoints/ServiceCollectionExtensions.cs
@@ -268,29 +268,6 @@ public static class ServiceCollectionExtensions
     }
 
     /// <summary>
-    /// Enables support for the Implicit and Hybrid Flows by registering the response-type
-    /// processors that emit access tokens and ID tokens directly from the authorization endpoint.
-    /// </summary>
-    /// <remarks>
-    /// OAuth 2.1 §1.4 deprecates the Implicit Grant. By default this library does not register
-    /// the <c>token</c> or <c>id_token</c> response-type processors, so the authorization endpoint
-    /// rejects requests for those response types with <c>unsupported_response_type</c> and the
-    /// discovery document advertises only <c>code</c> in <c>response_types_supported</c> and
-    /// omits <c>implicit</c> from <c>grant_types_supported</c>. Hosts that still need Implicit
-    /// or Hybrid (legacy SPAs, transition deployments) make a deliberate decision to opt in via
-    /// this method, mirroring the <c>EnablePasswordGrant</c> precedent for ROPC.
-    /// </remarks>
-    /// <param name="services">The <see cref="IServiceCollection"/> to register the Implicit /
-    /// Hybrid response-type processors in.</param>
-    /// <returns>The <see cref="IServiceCollection"/> so additional calls can be chained.</returns>
-    public static IServiceCollection EnableImplicitFlow(this IServiceCollection services)
-    {
-        services.TryAddEnumerable(ServiceDescriptor.Singleton<IAuthorizationResponseProcessor, TokenProcessor>());
-        services.TryAddEnumerable(ServiceDescriptor.Singleton<IAuthorizationResponseProcessor, IdTokenProcessor>());
-        return services;
-    }
-
-    /// <summary>
     /// Registers services required for JWT Bearer grant type, including JWT Bearer issuer provider,
     /// JWT replay prevention cache, and keyed caching decorator for JWKS fetching.
     /// </summary>
